implemented applyParameters in CdkChemical2FactoryImpl
